Fix the azure KMS provider
There were two issues: double hashing and the wrong signature formats.

We already hash the contents in SignMessage using a configurable
hash function, so we don't need this one here.

The other issue was caused by azure returning and expecting a raw byte sequence
of R||S concatenated, rather than an ASN.1 sequence. Big thanks to mwielgoszewski
for providing the tips and POC.

Signed-off-by: Dan Lorenc <lorenc.d@gmail.com>