Fix the unit tests with expired TUF metadata. (#1270)
These tests worked by mocking at the "isExpired" level. When the real files
ARE expired, but we mock them to be NOT expired, the code continues down a
path it shouldn't and fails later, trying to use expired metadata.

We should fix this "better" by generating real expired and unexpired metadata,
or changing the system clock somehow.

Signed-off-by: Dan Lorenc <lorenc.d@gmail.com>
Signed-off-by: Jason Hall <jason@chainguard.dev>