KMS: Change how the Azure authentication method is handled
This patch removes the requirement of having the environment variables
'AZURE_TENANT_ID', 'AZURE_CLIENT_ID' and 'AZURE_CLIENT_SECRET' set to
use the Azure KMS (KeyVault).

By removing the requirement, we enable usage of MSI (Managed Service
Identity) through the NewAuthorizerFromEnvironment() and we at the same
time add support for the NewAuthorizerFromCLI().

By splitting the function of calculating what method to use to it's own
function , getAuthenticationMethod(), it's possible to test the logic
separately.

We also introduce the new environment variable 'AZURE_AUTH_METHOD' which
if set to 'environment' will use FromEnvironment (may be useful for the MSI
case) and if set to 'cli' will use FromCLI.

If nothing is defined, FromEnvironment will be tested first and then
FromCLI(). :^)

Signed-off-by: Simon Gottschlag <simon.gottschlag@xenit.se>