Drop SHA1, SHA224 for RSA-PSS/PKCS#1, enforce for RSA-PKCS#1 (#234)
This sets the set of supported hashes to SHA256, SHA384,
and SHA512. SHA1 should not be used for RSA-PKCS#1v1.5,
as it is not collision resistant. Not an expert on this,
but technically SHA1 is acceptable for RSA-PSS due to how
the hash function is used, as part of its
"mask generation function". However, since SHA1 is not
supported by cloud KMS services, we should just remove it entirely.
SHA224 is also not supported anywhere.

RSA-PKCS#1v1.5 was not enforcing the set of hashes on load.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>