Skip strict check on PKCE discovery claim on Azure
This skips the check for the code_challenge_methods_supported claim
within the OIDC discovery document when configured to connect to an
Azure-based OIDC provider. AFAIK, Azure OIDC does in fact support PKCE
but does not advertise this in its discovery document so this should
enable additional use cases with cosign and Azure.

Partially-Fixes: https://github.com/sigstore/cosign/issues/1350

Signed-off-by: Bob Callaway <bob.callaway@gmail.com>