Make tag,key_ring,key optional.
This is make it possible to cut releases completely automatically.

The tag will now increment the patch value by one if the release tag is
not specified.
For example:
- 1.5.9 => 1.5.10

Keyring and Key value from GCP project are now hardcoded.

Will remove the input key_ring and key when it dependent workflows are
updated to no longer specify them.

Signed-off-by: Kenny Leung <kleung@chainguard.dev>