tuf: fix on-disk cache when writing targets in subfolders (#729)
Added a test, patched cosign locally, and can confirm this now works correctly with the v5 root changes.

Note: no old clients will break from this change. It's forward compatible.

It's in anticipation of a root update that we'll push to a separate repository as to not break clients who are still using this version. Clients won't break until the current repository loses lifetime: in late January next year.

Signed-off-by: Asra Ali <asraa@google.com>