fix: set go module directive to 1.22.0
The go directive sets the minimum version of Go required to use this
module. There's no need to set this to the latest semver patch version
of a given Go release unless the semantics of that version of Go are
required to build/test/use the module.

Notably since PR #1865 any consumers of sigstore as a library would need
to bump their own go directive to 1.22.8 in order to pickup the sigstore
1.8.10 release.

Ref: https://go.dev/ref/mod#go-mod-file-go

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>