add initial plugin support for KMSs  (#1901)
We're adding plugin support the sigstore to support separately-compiled KMS providers. We experimented with the hashicorp/go-plugin system suggested in #1658, and we would like to try kubectl's style of supporting plugin programs. It basically works like this.

The plugin is a separate program on the system's PATH, invoked as a command
the SignerVerifier interface arguments are sent as CLI arguments and file contents over stdin
the SignerVerifier interface return values are sent over stdout
You should be able to implement a plugin in any language, even shellscript. If using go, once you implement the SignerVerifier interface methods, your plugin program could be as simple as the example included in this PR at "test/cliplugin/localkms/main.go".

Signed-off-by: Ramon Petgrave <ramon.petgrave64@gmail.com>