Get algorithm details by OID
This will be used by Rekor to determine which hash function should be
used for a given key algorithm. We already have a function to map
PublicKeyDetails to AlgorithmDetails which works for keys, since Rekor
v2 requests will require the user specify this value. However for
certificates, we don't have this enum, so instead we'll use the OID from
the certificate.

We add a few additional options to differentiate between different RSA
key sizes, since RSA-PSS and RSA-PKCS#1v1.5 OIDs are the same for all
key sizes, and between Ed25519 and Ed25519-ph.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>