Add back ValidatePubKey to cryptoutils package
We moved this function to a dedicated package since the packages we
depended on from letsencrypt/boulder was large. This was a breaking API
change and ended up being a bit of a pain to update all Sigstore
clients. While this is done, we want to avoid breaking anyone else, so
we've removed the dependency on boulder by copying in the one file we
need and adding the previous function back to cryptoutils.

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>